attempt 3

+++ include/task2.h
@@ -3,7 +3,7 @@
 
 #include<math.h>
 
-bool checkPrime(unsigned long long value);
+bool checkPrime(unsigned int value);
 unsigned long long nPrime(unsigned n);
 unsigned long long nextPrime(unsigned long long value);
 #endif // !TASK2_H
