+++ src/task2.cpp
@@ -1,7 +1,7 @@
 #include "task2.h"
 
 
-bool checkPrime(unsigned long long value) {
+bool checkPrime(unsigned int value) {
 	for (unsigned int i = 2; i <= sqrt(value); i++) {
 		if (value % i == 0) return false;
 	}
